+++ src/enums.ts
@@ -1,5 +1,7 @@
-
-export enum EnumConfigs {
+export enum Commands {
+   InlineFoldToggle = "inlineFold.toggle",
+}
+export enum Configs {
    identifier = "inlineFold",
    regex = "regex",
    regexFlags = "regexFlags",