+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ## Change log:
+
+### 0.1.0
+    - Added offsets to the visible part of the editor.
+    - Registered toggle command (#1)
+    - Changed default regex to support react.
+    - Unfold selected lines.
+    - Fixed: multi cursor selection doesn't unfold text (#6)
+    - Support for react out of the box (#9) (#10) (#12)
+    - Fixed support for files with lines lower than visiual range
+
 ### 0.0.8
     - perf: ⚡ New approach to boost performance #2